3_SmartFeed/requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
gcloud==0.17.0
googleapis-common-protos==1.51.0
httplib2==0.18.1
idna==2.9
jieba3k==0.35.1
joblib==0.15.1
jws==0.1.3
lxml==4.5.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.18.4
oauth2client==3.0.0
pandas==1.0.3
Pillow==7.1.2
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-http-client==3.2.7
python-jwt==2.0.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.11.1
requests-file==1.5.1
requests-toolbelt==0.7.0
rsa==4.0
schedule==0.6.0
scikit-learn==0.23.1
scipy==1.4.1
sendgrid==6.3.1
six==1.15.0
sklearn==0.0
soupsieve==2.0.1
threadpoolctl==2.0.0
tinysegmenter==0.3
tldextract==2.2.2
tqdm==4.46.0
wincertstore==0.2
